+++ src/page/__Photo/PhotoList.tsx
@@ -15,132 +15,7 @@ import { useState } from 'react';
 import { P, match } from 'ts-pattern';
 import { RenderInView } from './RenderInView';
 import { VRChatWorldJoinDataView } from './VRChatJoinDataView';
-import { useComponentWidth } from './hooks';
-
-// 各セクションの領域の高さ、写真の幅、高さを計算するhook
-const usePhotoArea = (props: {
-  componentWidth: number | undefined;
-  gapWidth: number;
-}): {
-  len: number;
-  countByYearMonthList: {
-    photoTakenYear: number;
-    photoTakenMonth: number;
-    photoCount: number;
-    areaHeight: number;
-    columnCount: number;
-    rowCount: number;
-    photoWidth: number;
-  }[];
-} | null => {
-  const { data: countByYearMonthList } =
-    trpcReact.vrchatPhoto.getCountByYearMonthList.useQuery();
-
-  const getPhotoWidth = (componentWidth: number, gapWidth: number) => {
-    // 写真のサイズもコンポーネントの横幅によって動的に決定する
-    const photoWidthMax = match(componentWidth)
-      .with(P.number.lte(400), () => 100)
-      .with(P.number.lte(800), () => 150)
-      .otherwise(() => 200);
-
-    // そのままおいたときに最大何枚写真が並ぶか
-    const columnCountMax = Math.floor(
-      componentWidth / (photoWidthMax + gapWidth),
-    );
-
-    // カラム数が0にならないようにエラーハンドリング
-    if (columnCountMax === 0) {
-      return photoWidthMax;
-    }
-
-    // 残りの幅
-    const restWidth =
-      componentWidth - columnCountMax * (photoWidthMax + gapWidth);
-
-    // 残りの幅を埋めるために必要な写真の幅
-    const photoWidthAdditional = restWidth / columnCountMax;
-
-    // 合計幅がcomponentWidthになるように調整された写真の幅
-    let adjustedPhotoWidth = photoWidthMax + photoWidthAdditional;
-
-    // 調整後の写真幅を使用して計算された合計幅
-    const totalWidth = columnCountMax * (adjustedPhotoWidth + gapWidth);
-
-    // 調整後の写真幅が合計幅と一致しない場合、調整を再計算
-    if (totalWidth !== componentWidth) {
-      adjustedPhotoWidth =
-        (componentWidth - gapWidth * (columnCountMax - 1)) / columnCountMax;
-    }
-
-    return adjustedPhotoWidth;
-  };
-
-  const photoWidth = props.componentWidth
-    ? getPhotoWidth(props.componentWidth, props.gapWidth)
-    : 100;
-
-  const [resultByYearMonth, setResultByYearMonth] = useState<{
-    len: number;
-    countByYearMonthList: {
-      photoTakenYear: number;
-      photoTakenMonth: number;
-      photoCount: number;
-      areaHeight: number;
-      columnCount: number;
-      rowCount: number;
-      photoWidth: number;
-    }[];
-  } | null>(null);
-
-  const result: {
-    photoTakenYear: number;
-    photoTakenMonth: number;
-    photoCount: number;
-    areaHeight: number;
-    columnCount: number;
-    rowCount: number;
-    photoWidth: number;
-  }[] = [];
-
-  useMemo(() => {
-    if (countByYearMonthList === undefined) {
-      setResultByYearMonth(null);
-      return null;
-    }
-    if (props.componentWidth === undefined) {
-      setResultByYearMonth(null);
-      return null;
-    }
-    setResultByYearMonth({
-      len: countByYearMonthList.length,
-      countByYearMonthList: result,
-    });
-  }, [countByYearMonthList, props.componentWidth, photoWidth]);
-
-  console.log(`componentWidth: ${props.componentWidth}`);
-
-  if (countByYearMonthList === undefined) {
-    return null;
-  }
-  if (props.componentWidth === undefined) {
-    return null;
-  }
-
-  for (const countByYearMonth of countByYearMonthList) {
-    const columnCount = Math.floor(props.componentWidth / photoWidth);
-    const rowCount = Math.ceil(countByYearMonth.photoCount / columnCount);
-    const areaHeight = rowCount * photoWidth + (rowCount - 1) * props.gapWidth;
-    result.push({
-      ...countByYearMonth,
-      areaHeight,
-      columnCount,
-      rowCount,
-      photoWidth,
-    });
-  }
-  console.log(resultByYearMonth);
-  return resultByYearMonth;
-};
+import * as hooks from './hooks';
 
 const PhotoList = (props: {
   onSelectPhotoFileName: (fileName: string) => void;
@@ -158,85 +33,89 @@ const PhotoList = (props: {
 }) => {
   console.log('PhotoList');
   console.log(props);
-  const currentRef = React.useRef<HTMLDivElement | null>(null);
+  const parentRef = React.useRef<HTMLDivElement | null>(null);
   const rowVirtualizer = useVirtualizer({
     count: props.len,
-    getScrollElement: () => currentRef.current,
+    getScrollElement: () => parentRef.current,
     estimateSize: (index) => props.countByYearMonthList[index].areaHeight + 56,
-    overscan: 2,
+    overscan: 0,
   });
 
   useMemo(() => {
     // countByYearMonthList の変更をトリガーに再レンダリングを促す
+    console.log('measure');
     rowVirtualizer.measure();
   }, [props.countByYearMonthList]);
 
   return (
-    <div
-      ref={currentRef}
-      style={{
-        height: `${rowVirtualizer.getTotalSize()}px`,
-        width: '100%',
-        position: 'absolute',
-      }}
-    >
-      {rowVirtualizer.getVirtualItems().map((virtualRow) => (
-        <div
-          key={virtualRow.key}
-          style={{
-            position: 'absolute',
-            top: 0,
-            left: 0,
-            width: '100%',
-            height: `${virtualRow.size}px`,
-            transform: `translateY(${virtualRow.start}px)`,
-            justifyContent: 'flex-start',
-          }}
-          className="flex-wrap flex"
-        >
-          {(() => {
-            const countByYearMonth =
-              props.countByYearMonthList[virtualRow.index];
-            const {
-              photoTakenYear,
-              photoTakenMonth,
-              photoCount,
-              photoWidth,
-              rowCount,
-            } = countByYearMonth;
-            return (
-              <div
-                key={`${photoTakenYear}-${photoTakenMonth}`}
-                className="flex flex-col w-full"
-                style={{
-                  height: `${rowCount * photoWidth}px`,
-                }}
-              >
-                <div className="flex items-center justify-between px-3 py-2">
-                  <div className="flex items-center space-x-2">
-                    <Globe size={20} />
-                    <Badge>{`${photoTakenYear}年${photoTakenMonth}月`}</Badge>
+    <div ref={parentRef} className="h-full overflow-y-auto">
+      <div
+        style={{
+          height: `${rowVirtualizer.getTotalSize()}px`,
+          width: '100%',
+          position: 'relative',
+        }}
+      >
+        {rowVirtualizer.getVirtualItems().map((virtualRow) => (
+          <div
+            key={virtualRow.key}
+            style={{
+              position: 'absolute',
+              top: 0,
+              left: 0,
+              width: '100%',
+              height: `${virtualRow.size}px`,
+              transform: `translateY(${virtualRow.start}px)`,
+              justifyContent: 'flex-start',
+            }}
+            className="flex-wrap flex"
+          >
+            {(() => {
+              const countByYearMonth =
+                props.countByYearMonthList[virtualRow.index];
+              const {
+                photoTakenYear,
+                photoTakenMonth,
+                photoCount,
+                photoWidth,
+                rowCount,
+              } = countByYearMonth;
+              return (
+                <div
+                  key={`${photoTakenYear}-${photoTakenMonth}`}
+                  className="flex flex-col w-full"
+                  style={{
+                    height: `${rowCount * photoWidth}px`,
+                  }}
+                >
+                  <div className="flex items-center justify-between px-3 py-2">
+                    <div className="flex items-center space-x-2">
+                      <Globe size={20} />
+                      <Badge>{`${photoTakenYear}年${photoTakenMonth}月`}</Badge>
+                    </div>
+                    <div className="flex items-center space-x-2">
+                      <Image size={20} />
+                      <Badge>{`${photoCount}枚`}</Badge>
+                    </div>
                   </div>
-                  <div className="flex items-center space-x-2">
-                    <Image size={20} />
-                    <Badge>{`${photoCount}枚`}</Badge>
-                  </div>
+                  <PhotoListYearMonth
+                    onSelectPhotoFileName={props.onSelectPhotoFileName}
+                    photoTakenYear={photoTakenYear}
+                    photoTakenMonth={photoTakenMonth}
+                    photoWidth={photoWidth}
+                    gapWidth={props.gapWidth}
+                  />
                 </div>
-                <PhotoListYearMonth
-                  onSelectPhotoFileName={props.onSelectPhotoFileName}
-                  photoTakenYear={photoTakenYear}
-                  photoTakenMonth={photoTakenMonth}
-                  photoWidth={photoWidth}
-                  gapWidth={props.gapWidth}
-                />
-              </div>
-            );
-          })()}
-        </div>
-      ))}
+              );
+            })()}
+          </div>
+        ))}
+      </div>
     </div>
   );
 };
+
+PhotoList.whyDidYouRender = true;
 
 const PhotoByPathRevalidateOnPathNotFound = (props: {
   photoPath: string;
@@ -312,15 +191,16 @@ export const PhotoListAll = (props: {
   // component の横幅
   const currentRef = React.useRef<HTMLDivElement | null>(null);
   // component の横幅
-  const componentWidth = useComponentWidth(currentRef);
+  const componentWidth = hooks.useComponentWidth(currentRef);
   console.log(`componentWidth: ${componentWidth}`);
   const gapWidth = 4;
-  const photoAreaList = usePhotoArea({
+  const photoAreaList = hooks.usePhotoArea({
     componentWidth,
     gapWidth,
   });
 
   const renderContent = () => {
+    console.log('renderContent');
     if (photoAreaList === null) {
       return null;
     }
@@ -336,10 +216,8 @@ export const PhotoListAll = (props: {
   };
 
   return (
-    <div className="h-full w-full relative">
-      <ScrollArea className="h-full absolute overflow-y-auto" ref={currentRef}>
-        {renderContent()}
-      </ScrollArea>
+    <div className="h-full w-full relative" ref={currentRef}>
+      {renderContent()}
     </div>
   );
 };
